+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added: `ignorePath` (for JS) and `--ignore-path` (for CLI) options.
 - Added: `at-rule-name-newline-after` rule.
+- Fixed: `function-whitespace-after` ignores `postcss-simple-vars`-style interpolation.
 - Fixed: `function-url-quotes` ignores values containing `$sass` and `@less` variables.
 - Fixed: selector-targeting rules ignore Less mixins and extends.